Don't translate a window's location if its parent is the root.
Previously the locations of windows were translated under the
assumption that their parent was not the root.  In some window
managers, e.g. xmonad or dwm, many windows do have the root as a
parent.  In this case the location returned by "getwindowgeometry" was
wrong.  Now xdotool will check that the parent's window is not root
before doing the translation.