Set display_name field
Caught by clang static analyzer: Dead assignment.

`display_name` does not seem to be used anywhere, so it could actually
be removed... but the xdo.h header is public, so removing the field
would break the interface.

For now, just set the `display_name` field, but do not attempt to free
static memory.