xdo: fix undef behavior in xdo_wait_for_window_size
Caught by clang static analyzer: Result of operation is garbage or
undefined.

This prevents the function from trying getting the window size on
something that is gone (BadWindow). The caller (cmd_windowsize) does not
care about the status code, but that probably does not matter as it is
used for synchronization and technically "the window is not of the old
size" anymore.