xdo_search: fix memleaks, pre-compile regexes optimization
If no regex is given, a default one matching none would be used. This
memory was not freed. Now it is freed and besides that, all regexes
are now pre-compiled before iterating through all windows.

(as a side-effect of the optimization: return early when a regex fails
to compile).

Other changes: constantify regex_t* where applicable, whitespace
cleanup.