Add a test for search race condition
When window disappear while search is trying to access it (either check
its properties, or enumerate its children), X server will
generate BadWindow error. Try to trigger this race, by opening 100
short living windows and execute xdotool search at the same time. The
search should not crash, and since the window subject to search is there
all the time, it should be found.

The test relies on a specific siming, specifically that launching 100
instance of xterm takes more than 100ms, and that going from setup to
the actual test takes less than 100ms. Those conditions are rather
reasonable, and the test works reliably for me (at this stage - it fails
- the fix is comming in the next commit).

Related to #60