Fix race condition in search command
When a window is destroyed while search command tries to inspect it (get
its title, class or so), the X server generates BadWindow error. The
default error handler terminates the program. This means the search can
fail if any window is destroyed in the meantime, regardless if the
window matches the search criteria or not.

Fix this by setting an alternative error handler for the search time,
that ignores BadWindow error. This makes the relevant libX11 functions
return (with a negative status) in case of BadWindow error, instead of
interrupting the whole xdotool process.

While at it, fix uninitialized 'children' variable in XQueryTree error
handling.

Fix #60