Detect if the X server is XWayland and report an error.
Also: Detect if $DISPLAY env is not set and report a more informative
error message.

For the purposes of libxdo/xdotool, XWayland does not work, so it's best
to notify the user accordingly if detected.

This adds a dependency on XInput `libxi` because the only[1] way I have
found to detect XWayland is through this extension

[1] There is another detection method which uses
  XFree86-VidModeExtension, but I don't know how common that extension is
  compared to XInput2.

Fixes #341.
Related to #337.