Typo: move a window is not the business of get_desktop_for_window
get_desktop_for_window description has the exact same sentence as set_desktop_for_window description. Since moving is not the business of a get command, this is most probably some sort of an extension to the wrong place. But I have not checked the actual C code. Perhaps this is a more fundamental error.