Remove unused dependency on libXi
In commit 6c1db901e22f21eb8ffa1ea03da562a59c115067, the
appears_to_be_wayland function was reverted (commented out, really), but
the includes and scaffolding for it were left in. This has a couple of
effects:

- The compiler emits a warning for unused functions
- An additional header include is used
- The link command is longer, and tries to link to libXi, resulting in
  overlinking unless -Wl,--as-needed is used

The original rationale for keeping it:

> I've left the "is on wayland?" detection code that is otherwise now
> unused, in case we want it later.

The code has gone untouched for several years now, so it is probably
time to remove it altogether. In a pinch, it can always be retrieved via
`git log`.