At shift level one, explicitly check the keycode with no modifiers.
Some keys have multiple mappings for the same shift level *and* omitted
mappings for that level.

For example, there's an omitted mapping for F1 on my keyboard such that
keycode 67, with no modifiers, produces F1. There's also a mapping for
shift level one where keycode 67 with Mod1Mask also produces F1.

However, most programs receive this as "Alt + F1" and not "F1"

So, just directly check "Does this keycode with no modifiers produce a
keysym at this shift level?" and use that if one is found.